Refactor typings for guild.ts
