+++ src/commands/milestone.ts
@@ -1,5 +1,5 @@
 import { Message } from "discord.js";
-import { guild_db, milestone_t } from "../database/guild";
+import { Guild, milestoneT } from "../@types/guild";
 import { execute_args, metadata_t } from "../handlers/commands";
 import { isNumber, embed } from "../utils";
 
@@ -17,7 +17,7 @@ type milestone_function_t = (
   message: Message,
   count: string,
   name: string,
-  gdb: guild_db,
+  gdb: Guild,
 ) => object;
 
 enum ErrorKind {
@@ -97,9 +97,9 @@ function listMilestones(
   message: Message,
   _count: unknown,
   _name: unknown,
-  gdb: guild_db,
+  gdb: Guild,
 ): object {
-  const milestones = gdb.get("milestones") as milestone_t;
+  const milestones = gdb.get("milestones") as milestoneT;
 
   if (Object.keys(milestones).length === 0) {
     throw ErrorKind.NoMilestones;
@@ -126,7 +126,7 @@ function addMilestone(
   _message: unknown,
   count: string,
   name: string,
-  gdb: guild_db,
+  gdb: Guild,
 ): object {
   if (!isNumber(count)) {
     throw ErrorKind.InvalidNumber;
@@ -154,7 +154,7 @@ function removeMilestone(
   _message: Message,
   count: string,
   _name: unknown,
-  gdb: guild_db,
+  gdb: Guild,
 ): object {
   if (!isNumber(count)) {
     throw ErrorKind.InvalidNumber;
