+++ src/handlers/commands.ts
@@ -1,10 +1,10 @@
 import { Message } from "discord.js";
-import { guild_db } from "../database/guild";
+import { Guild } from "../@types/guild";
 import { commands } from "../commands";
 
 export type execute_args = {
   message: Message;
-  gdb: guild_db;
+  gdb: Guild;
   args?: string[];
 };
 
@@ -27,7 +27,7 @@ export type command_t = {
 
 export async function commandHandler(
   message: Message,
-  gdb: guild_db,
+  gdb: Guild,
 ): Promise<void | Message> {
   // Split message into arguments (delimited by spaces in the message).
   const args = message.content.slice(gdb.prefix.length).trim().split(/ +/);
